+++ Sphinx/ChangeLog.rst
@@ -2,7 +2,7 @@ ChangeLog
 =========
 * May 25, 2014
 
-  - Added cache for colors.
+  - Added cache for colors and option ``--print-stats-cache``.
   - First public version of ``.hgignore`` and ``Makefile``.
   - Off the mixer if no sound is loaded.
   - Updated ``example/RiceRocks_Asteroids.py``.